[No-jira] 낙관락과 Spring Retry를 통한 리뷰 좋아요 동시성 제어 (#208)
* feat: 리뷰 엔티티와 도메인에 @version 컬럼 추가

* feat: 리뷰 엔티티에 likesCount 증감 메서드 추가

* feat: 리뷰 레포지토리에 낙관락 기반으로 좋아요 수 업데이트 구현

* feat: 리뷰엔티티 version 컬럼 디폴트 0L 설정

* feat: 리뷰 좋아요 토글 메서드에 낙관락 및 재시도 로직 구현

* test: 멀티스레드 낙관락 정합성 테스트

* feat: setter로 hibernate 변경 감지 가능하도록 변경

* feat: spring retry 적용헤 재시도 로직 구현

+++ usecase/src/main/java/org/depromeet/spot/usecase/service/review/like/ReviewLikeService.java
@@ -1,18 +1,25 @@
 package org.depromeet.spot.usecase.service.review.like;
 
-import org.depromeet.spot.common.annotation.DistributedLock;
-import org.depromeet.spot.domain.review.Review;
+import java.util.ConcurrentModificationException;
+
 import org.depromeet.spot.domain.review.like.ReviewLike;
 import org.depromeet.spot.usecase.port.in.review.ReadReviewUsecase;
 import org.depromeet.spot.usecase.port.in.review.UpdateReviewUsecase;
 import org.depromeet.spot.usecase.port.in.review.like.ReviewLikeUsecase;
 import org.depromeet.spot.usecase.port.out.review.ReviewLikeRepository;
+import org.depromeet.spot.usecase.port.out.review.ReviewRepository;
+import org.springframework.orm.ObjectOptimisticLockingFailureException;
+import org.springframework.retry.annotation.Backoff;
+import org.springframework.retry.annotation.Recover;
+import org.springframework.retry.annotation.Retryable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import lombok.Builder;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 
+@Slf4j
 @Service
 @Builder
 @Transactional
@@ -22,31 +29,38 @@ public class ReviewLikeService implements ReviewLikeUsecase {
     private final ReadReviewUsecase readReviewUsecase;
     private final UpdateReviewUsecase updateReviewUsecase;
     private final ReviewLikeRepository reviewLikeRepository;
+    private final ReviewRepository reviewRepository;
 
     @Override
-    @DistributedLock(key = "#reviewId")
+    @Transactional
+    @Retryable(
+            retryFor = ObjectOptimisticLockingFailureException.class,
+            maxAttempts = 4,
+            backoff = @Backoff(delay = 100))
     public boolean toggleLike(final Long memberId, final long reviewId) {
-        Review review = readReviewUsecase.findById(reviewId);
+        boolean exists = reviewLikeRepository.existsBy(memberId, reviewId);
 
-        if (reviewLikeRepository.existsBy(memberId, reviewId)) {
-            cancelLike(memberId, reviewId, review);
+        if (exists) {
+            reviewLikeRepository.deleteBy(memberId, reviewId);
+            reviewRepository.updateLikesCount(reviewId, false);
             return false;
         }
 
-        addLike(memberId, reviewId, review);
+        ReviewLike like = ReviewLike.builder().memberId(memberId).reviewId(reviewId).build();
+        reviewLikeRepository.save(like);
+        reviewRepository.updateLikesCount(reviewId, true);
         return true;
     }
 
-    public void cancelLike(final long memberId, final long reviewId, Review review) {
-        reviewLikeRepository.deleteBy(memberId, reviewId);
-        review.cancelLike();
-        updateReviewUsecase.updateLikesCount(review);
-    }
-
-    public void addLike(final long memberId, final long reviewId, Review review) {
-        ReviewLike like = ReviewLike.builder().memberId(memberId).reviewId(reviewId).build();
-        reviewLikeRepository.save(like);
-        review.addLike();
-        updateReviewUsecase.updateLikesCount(review);
+    // 예외 처리를 위한 Recovery 메소드
+    @Recover
+    public boolean recoverToggleLike(
+            ObjectOptimisticLockingFailureException e, Long memberId, long reviewId) {
+        log.error(
+                "Failed to toggle like after 3 attempts for memberId: {} and reviewId: {}",
+                memberId,
+                reviewId,
+                e);
+        throw new ConcurrentModificationException("좋아요 처리 중 충돌이 발생했습니다. 잠시 후 다시 시도해주세요.");
     }
 }
